Make lock refresh + failure cache configurable #543
Make lock failure cache configurable,
the defaults should be as they're today.
Now failure cache expiry time value would
be picked from ecc.yml.

Lock refresh is dynamic based on the TTL
of ecchronos.lock table

+++ core/src/main/java/com/ericsson/bss/cassandra/ecchronos/core/CASLockFactoryCacheContext.java
@@ -54,10 +54,10 @@ public final class CASLockFactoryCacheContext
     public static class Builder
     {
         private LockCache myLockCache;
-        private long myLockUpdateTimeInSeconds;
+        private int myLockUpdateTimeInSeconds;
         private int myFailedLockRetryAttempts;
 
-        public final Builder withLockUpdateTimeInSeconds(final long lockTimeInSeconds)
+        public final Builder withLockUpdateTimeInSeconds(final int lockTimeInSeconds)
         {
             myLockUpdateTimeInSeconds = lockTimeInSeconds;
             return this;
